Add tests for thread affinities of Safe AsyncTask
This ensures that methods that shouls be running on the UI thread
are running on the UI Thread, and methods that should not be are
not running on the UI Thread.

The Test Cases class now inherits from ActivityUnitTestCase because
AsyncTask needs to be started from the UI thread, if not the
onPostExecute method is not run. Known semi-bug.

Change-Id: I816e8ba9cc77c3a36ac17b6eac18fa2d9ab03152