Do not keep all of past page fragments fully in memory
We're still saving their state, so the pages will come right back
up. We might want to discard them after some depth to preserve
memory, but that can come later.

This means we lose the scroll position when you hit back, but
that needs to be solved anyway in the 'fragment re-created'
case for supporting rotates, etc - so will get to it eventually.

Change-Id: I0f4368bf9cd2b6d5365e1c8db28a3c140c07863f