Add a level of indirection when sending messages to WebView
Since apparently just loading a javascript: URL is a blocking
operation and also quite slow when passing in a large amount of
data, we just pass in a pointer and then use another method to
retreive the data.

The 'other' method is using the JavascriptInterface. This is
known to fail in some 2.3 devices with low memory and also
in 2.3 emulators. But the performance boost that we get
from this is significant enough to risk it, and I think
in the future if necessary we can implement a fallback in
the cases when this method of communication does not
work.

Change-Id: I16c10fb1faea9bec1d6a69ac53d90ac88653fcb1