Handle out of order network responses to search query
If a previous search query returns later than a newer search
query, it will confuse the reader by displaying the results of the
older search query, overriding the newer one.

Cancelling the older search task will not actually cancel the
HTTP request, but that's okay. It does prevent the onFinish
and onCatch from firing, which should be good enough for this
purpose.

Change-Id: I0c9e1204f7fbddf00a5077c33c725dc22da9ae9b