[Gingerbread] Disable crashy AutoCompleteTextView in edit summary
Leaving it on for ICS and later, just disabling the autocomplete
on Gingerbread by not setting the adapter.

For some reason the autocomplete popup view crashes on
Gingerbread. This seems to be related to styles but I
can't quite figure out why.

This call in AutoCompleteTextView.buildDropDown ends up failing:

            mDropDownList.setSelector(mDropDownListHighlight);

because mDropDownListHighlight seems to be null instead
of an expected drawable, and that ends up failing when used.

But the styles *look* ok, and I don't see any missing drawables
related to that style...

Change-Id: I96dfdb3b63adca6ccc4bf043d03c8224905fbf23