Always use bundled styles until June 17th UTC
MobileFrontend changes that we depend on have not been deployed
yet. They are in master, and not even in 1.24wmf7. As a result
we get older styling for anything past the first page.

These changes will be in 1.24wmf8, which will be branched
on 5th June on to TestWiki, and will be on enwiki on
June 12th (according to
https://wikitech.wikimedia.org/wiki/Deployments).
Setting this to 17th gves us a few more days to fix things
in the cluster before it goes live, if issues exist.

This removes the ability to modify the styles in place
for the first two weeks of the beta app release, but
I think that is better than broken styles or trying
to SWAT the MobileFrontend changes.

Change-Id: I91205fe2d10c50406d50c7a1c6bbe0f547fd5940