Save images of saved page
Images are stored in the same directory as the saved page.
No extra code for tracking needed, and better for consistency
vs. the database approach.
Another plus is that when we delete the saved page the
associated images are removed automatically.

Parsing of HTML is done using Html.fromHtml with a custom imageGetter.
This uses TagSoup internally, so no need to include an HTML parsing library.

When a saved page is loaded the img src URLs get replaced
by file URLs pointing to the saved image files.

Change-Id: I1c81ccefce50d5b4fe3c1790bfe803778315bd4c