Nearby
Get list of pages of nearby locations (up to 50 within 10 km
- the API doesn't let us increase the radius past 10km).
Shows list sorted by distance.

The distances update automatically but the list does not
update until the refresh button in the action bar is pressed.
This is done as to not confuse the user while moving or even
just scrolling around.

Change-Id: Ib9cb06b0d1e959431e1ed7188bc0e6ccf35c2bec