Swap the order of language names in language selectors.
In the "Read in another language" option and "Wikipedia language" setting, the
languages are displayed with the current device language's name in large text,
and the name of the language in the language itself in small text. This patch
flips that behaviour, so that the name of the language in the language itself
is prioritised.

Change-Id: I672cee2cf8b6e5ffdc25356d285b2235532c671d