Show ToC automatically; plus hints
Automatically opens the ToC (Table of Contents) drawer when
the user views a page with two or more sections until she
swipes to open the ToC or presses the Got it button.
Adds instructions on how to open the ToC by swiping horizontally.
I hope the translators get this right for RTL vs LTR.

The only way to get this to appear again after it was acknowledged
via "Got it" is to clear data for our app in the settings.

Change-Id: I8326e38ecc144a4e3acf53657657ee112361434a