Really mark ToC as known if Got it was pressed
There was a bug that it still showed the ToC intro when one
pressed on the Got it button but used the ToC action
bar button to open the ToC again.

Also adjusting the size of the hand swipe image to be smaller
for devices with smaller height, so that the Got it button fits
on the screen. Unfortunately, when orientation is changed
it doesn't update since we still using the orientation hack in the
AndroidManifest. So, when changing from portrait to landscape
there is still a chance that the button would be hidden but
switching back to portrait would fix this.

Change-Id: I229af1da44bc9b7e4ead67a653420ef0022b7127