Add explicit super keyword to log method of all classes extending Funnel.
All classes that extend Funnel have their own implementation of the log method
but it always ends up calling the super's log method. However, the code looks
confusing and doesn't make that obvious.

This patch adds the super keyword before every call to the parent's log
function, for clarity. Some classes already had the keyword in there, but this
patch adds it to the rest. The actual functionality before and after this patch
should be exactly the same.

Change-Id: I03a65d4a95500568b569e2a04f5814ccd179b40d