Show message to user when filtering interwiki links returns no results.
When the user goes to an article that has interwiki links, views the list of
links, filters the list, and gets no results, display a message to them
informing them that no languages matched their query.

Also add small left and right margins to the messages that the activity can
present, so that those messages wrap more cleanly and consistently.

Change-Id: I68fbb60c3e8210c83744ee2b5aa86c2041d2bf6d