Save nearby results between activity restores
When the Nearby activity resumes we still have the same
results, even if it was destroyed because of low memory.
This will make navigation easier since we don't automatically
change the list of results when the location changes.
If the user explicitly presses the refresh button or restarts
the activity then a new list of results will be retrieved.

NearbyResult is an object now, and parcelable.
Together with the locations (prev/next), it is stored in
a Bundle during onSaveInstanceSate;
to be used if the activity has to be recreated.

Change-Id: I3aba3f2202c9d2f08de88384ca450fee322dfc13