Add flag `FEATURE_ACTION_MODE_OVERLAY` for an overlaying actionmode
Without this flag the actionmode, when activated (through a long-click)
looks like [this](http://i.imgur.com/nPSD8WF.png).
With it it looks like [this](http://i.imgur.com/tBR6bG5.png).

It bugged me that the content of the screen got pushed down every
time I tried to copy something. With this it doesn't happen.

Change-Id: Id059e9f53c32785dd1cc97a7203b8dedee09dd90