Drag some Wikipedia domain stuff into the Site class
So it's (mostly) all in one place and changing some code to make it target a LAN wiki is easy.
This means you basically just have to change 5 methods in the Site class to get it to run on your
own wiki. (different API endpoint URL, no HTTPS, no external load.php, change of domain in forLang
and isSupportedSite)

Doesn't touch EventLogging or MobileApp stuff

Relies on I4b54b236

Change-Id: I8ee046f5f5b25b81db4b5b039503e8ec1e11ceb0