Move to Gradle build
This is not backwards-compatible with the regular Maven build!
Henceforth thou shalt use the Gradle build
instead of the old Maven build.

Gradle package (applicationId) rename:
Use ${applicationId} in AndroidManifest for providers.

Set Intent for AboutActivity in code vs preferences.xml
if a non-standard applicationId is used.

Uses AndroidManifest.xml snippets for every flavor
for the channel value, so no need to use the
prepare-release script for that and
for setting versionCode and versionName.

Added amazon flavor, removed master and ci.

Added new super-simple bump script which
increments the versionCode in the build.gradle file.

Added new make-release.py script which runs the
Gradle build and also provides a second step to
create and push the tags.

Removed the old scripts we don't use anymore.

Change-Id: I3eca0bf0d9e24fcf738455aedc64e095bc29b737