Anchor crouton position in Nearby
We're using Crouton library to display toasts when
the Nearby feature encounters errors.
Since we moved to an action bar that can be overlaid
those croutons were a bit too low, showing a gap
between the action bar and the top of the crouton.

Also added missing contentDescription in layout file.

Change-Id: I15b825f4933e425d7a0b5992fa1a1d270b2b6475