Only fire onboarding 'start' event the first time the activity is created.
Right now, the 'start' event for onboarding is fired every time the activity
is created. This means that if the user closes the app and leaves it in the
background for a long time (and then the activity is destroyed by Android),
the event is fired again when they reopen the app. This should track on a
user-by-user basis, so this is actually pointless.

This patch changes the activity so that it only fires the 'start' event the
first time the activity is created. That way, we can get closer to figuring
out the true bounce rate (on a user-by-user basis, rather than a view-by-view
basis).

Change-Id: I3f8f9e6290e3ab52856e5f39755b61fc6b8de04c