Add Wikidata descriptions to full text results
Switch to search generator so we can get the Wikidata IDs
(wikibase_item).

Once we have a list of wikidata IDs we haven't check
descriptions for, send off one request to wikidata to get
the descriptions for the given IDs.

Also updated the item layout for full text search
results to not cut off the descriptions.

Change-Id: I1e20a652a3e3b81301708ab60eeba416bd138af6