Refactor handling of shared preferences and feature install IDs.
This patch changes the shared preferences system of the app so that it's at
the correct level of abstraction. Presently, classes extending Funnel interact
directly with the preference manager to get feature install IDs, and that
should probably be handled centrally in WikipediaApp rather than in each
class.

This patch:
* Makes the SharedPreferences of the app a field in WikipediaApp.
* Instantiates the SharedPreferences field in the WikipediaApp constructor.
* Changes ad-hoc retrieval of the SharedPreferences to use the field instead.
* Creates generic retrieval method for install IDs in WikipediaApp.
* Creates accessor methods for the install IDs, which use the generic method.
* Changes classes that extend Funnel to use the install ID accessor methods.

Change-Id: Ic13b36ba1d4bcc766265f5944eb2f23c95869a8f