Change ActionBar search text every time a "Did you mean" query is run.
When the full text search API thinks you've made a typo, it returns a "Did you
mean" result, which is a suggestion for a new search. When you tap on that
result in the app, it runs a new full text search query with that term.

This patch makes it so that when such a "Did you mean" query is run, the text
in the search box is updated so that it matches the query that's run, so that
it's clearer to the user how the results match the query that was initiated.

Change-Id: Ib1195b9649e01c46425f1e86523559235c108902