Migrate all app install IDs to be the same as the old reading action ID.
As per the meeting between Product and Analytics, we're migrating all app
install IDs in all app schemas to be the same, so that the tables can be
joined and overlap in users between features can be found.

This patch changes all app install IDs so that they're the same as the install
ID that was previously used for reading actions. That way, the analysis that's
generated on an ongoing basis by analytics will remain unchanged, whereas the
EventLogging data install IDs will all change, but simply go in to new tables.

Change-Id: I747cc77afa63442c98f55ea2851dbd89940a2407