Make error messages in account creation slightly more prominent.
Right now, when a user tries to create an account with a username that is
already taken, the app simply displays a little red exclamation mark next to
the username field. When you tap on the username field, an error message is
displayed.

This patch makes it so that when you try the same thing, focus is given to the
username field so that the message displays immediately after the user has
made it, giving the user more feedback about what went wrong. It also does
this for every other kind of message which uses setError().

Bug: 72923
Change-Id: Ia5da8095e3b80463b6cd037bb931202c05369c74