Change width of tappable area for "Skip" in OnboardingActivity.
We're seeing a rather large drop off rate on the onboarding screen, and it was
pointed out that the tappable area of the "Skip" button is actually much
smaller than the other buttons because it's a text view.

This patch changes the tappable area of the skip button to be much wider so
that users are more easily able to skip onboarding without having to tap
exactly on the skip text.

Change-Id: I93e8c43a50029a7096c20a107b3591dbe392665f