Show confirmation dialogue when users click on links in preview.
Currently, if you tap on a link during preview, you leave your edit and go to
the page you clicked on. The mental model here is actually a bit unclear. In
any case, leaving your edit when you tap on a link is quite jarring, and the
fact that there's no warning before leaving makes it worse.

In absence of more tightly refining the mental model, this patch simply throws
up a confirmation dialogue when a user taps on a wikilink during preview, and
asks them whether they really want to leave the edit workflow or not.

Change-Id: I4e651ad7a548db17702a6165b44b6caf56c93d4e