Ensure Wikidata description is always shown in correct language in lead image
Right now, if you tap on "Read in other languages" on an article and choose
another language, the Wikidata description is displayed in the language you
just came from. This is because when the page in the other language is loaded,
the description is fetched in the primary language of the app and not in the
language of the actual page.

This patch:
* Refactors WikidataCache to allow you to load descriptions in a different
language from the primary language of the app, if you want to.
* Refactors WikidataCache to only actually cache results if the result is in
the language the app is set to, so that fetching a result in a different
language for a single page doesn't cause that description to appear in the
wrong language in places like search.
* Changes the lead image code to use this refactored code to fetch the
descriptions in the language of the article, rather than the default language.

Bug: 73598
Change-Id: If85a6f3d15a9758b1b8efc6faf96032893dc35df