Improve infobox/edit-pencil alignment.
The real issue was that some infoboxes were not properly moved to the
bottom of the lead paragraph. This patch moves any tables (including
infoboxes) inside the lead section to the bottom of the first paragraph,
which effectively leaves the edit pencil aligned with the first two lines
of the paragraph.

Change-Id: Iaf9ec6beec7bfd5bdcd00934371577a117475d00