Fix behavior of long-press in Nearby.
Before this patch, long-pressing an item in Nearby took the user to the
Google Maps app, *and* simultaneously navigated to the page that was
pressed (which could potentially cause a crash, as well).

This makes it so that long-pressing an item *only* takes the user to the
Maps app, and doesn't navigate to the selected page.

Change-Id: I96373bd3b4de4219af244f81bc522b94e77070a7