Wikidata descriptions retrieved via new PageTerms API
This patch affects retrieving descriptions for
- Full text search items
- Lead image
- Disambig items

Had to change the WikidataCache to use PageTitles instead of
Wikidata IDs as keys.

Always add the description of the clicked on item in full text search
to the cache so we don't even try to get it again when the page is opened.

Change-Id: I6b308f57de805b6d1193cf6b55e5d74307f1aaa0