Add hidePageContent convenience method to PageViewFragmentInternal
Currently, there are three separate places in PageViewFragmentInternal where
you want to hide all of the content of the page. Previously this was just the
web view, but now we have extra native components for content (lead images,
read more) which need to be hidden too.

This method adds a hidePageContent convenience method to the class, to reduce
code duplication.

Change-Id: Iaa6612b0aa521236b6a00c09f54f2676b468317d