Legal requirement: add links to license text for libraries
License links are added to credits in About screen.
License text is bundled with the app.
Ordered the libraries alphabetically.

Was originally planning on creating a script to update the
license texts but most are not in separate files.
We just try to remember to update the licenses when we update
the libraries.

Change-Id: I74350f1215c859d6a3f2275891c7a2e2660e01c6
(cherry picked from commit ec0f32778487199c96b29d073eb64b7e13b745e8)