Refactor code for storing preference for EventLogging opt-out
Right now, Funnel and MccMncStateHandler have to interact directly with the
SharedPreferences instance to get whether EventLogging is enabled or not. This
is quite clunky.

This patch abstracts this interaction to the WikipediaApp class, so that these
classes can use app.isEventLoggingEnabled() rather than interacting directly
with the SharedPreferences instance.

Change-Id: Ibba0e4baab923b4095cc8aa9f5a057f40f379977