Fix possible NPE(s) when accessing parent fragment.
In our LeadImagesHandler and BottomContentHandler, we use the
getFragment() function which returns the "internal" fragment object that
is currently hosted by the PageViewFragment.  The problem is that, when a
PageViewFragment is detached, it sets its instance of the internal object
to null.  So, if the Lead or Bottom components try to access the internal
fragment using getFragment() from the result of an AsyncTask after the
parent fragment is detached, they will get a null pointer.

This patch refactors things so that the internal object itself is passed
to the Lead and Bottom components, so that they will never receive a null
reference to it.

Change-Id: Ie78c76e12128a89218bcea8d65fbfeed01278a63