Add X-WMF-UUID to Wikipedia for Android requests
See also:
Ia0ab1691a31b9555721291803481104cb808689c

For now, the inclusion of the UUID in other places not to be removed.

On the Android it platform it seems that per-funnel identifiers
are part of the event data itself, and as with iOS, the UUID is
constructed in a reliable manner.

For now the goal is to ensure the goal is to ensure the presence
of the accurately set X-WMF-UUID header without causing other
unintended side effects in the app.

It should be possible to query request logs once this code and its
counterpart, in the following change

Ieeb3cc38e8d2f3244ec63d5345d4556e2f4a5df9

, are both deployed. If the values match when expected to match,
then it should be possible to remove the request path UUID in other
non-event payload contexts (e.g., PageViewFragmentInternal.java's
builder object with parameter "appInstallID").

Change-Id: I1f1cb4629cf7fe247f399a6d06feb7602ad0b455