Fix lead image click area in 2.3
The problem was that, even though the ViewHelper library (nineoldandroids)
correctly slides the lead image component, it does not successfully slide
the logical click area that corresponds to the visual position of the
component! So, the lead image remains "clickable" even after sliding away.

The solution is to stop using the ViewHelper library, and shift the
position of the lead image component using standard margins.

There was another slight issue, however:  in Android 2.3, using negative
margins is only possible for a component that has a LinearLayout as a
parent.  So, I had to put the lead image container into a dummy
LinearLayout, just to achieve negative margins.

Thanks again, 2.3!

NOTE: The same issue is currently present in the Search bar at the top,
which slides away in a similar fashion. This will be the subject of a
subsequent patch.

Bug: T84998

Change-Id: I79120601b156e151a8b9ddabcfef2952c1555de5