Get selected text from the WebView.
Sort-of hacky method that works by programmatically copying the selected
text to the clipboard, then peeking at the clipboard contents. (The actual
WebView provides no way of getting selected text)

Only works with API >10, which is fine since API 10 doesn't provide
long-press context functionality for the WebView anyway.

Also moved Bus-related methods into an inner class, because of this issue:
https://github.com/square/otto/issues/37

When the Bus registers a class for event notifications, it uses reflection
on all of the class's methods and fields. If a field does not exist in the
current API, it crashes.  Therefore, it's suggested to use an inner class
that only contains methods specific to the Bus.

Change-Id: I6c29eff4d9930bd218a9f686493cbcbc67d78691