Create image using text snippet
To create an image to be shared we create a new fixed size
Bitmap that is a bit taller than what we need.
We manually keep track of the vertical space (y) we draw on.
Once we're done drawing we adjust the height of the bitmap.

If we have a lead image, it goes on top,
followed by the title, an opening double quote sign,
then the text, and lastly the Wikipedia wordmark.

The lead image portion gets clipped and scaled to fit the width
of the screen. It's reusing the imageBaseYOffset for the face detection
adjustment and the Bitmap from LeadImageHandler.

The text goes through some minor regex replacements to get rid
of foot notes[1], remnants of IPAs, and spaces at beginning and end (trim).

If less than 6 characters remain then we don't do anything.
It is assumed that this rather arbitrary number is the absolute
minimum length of text in any language/character set to convey an idea/fact.

If the title or text is too long we first reduce the
font size. If that is not enough it gets ellipsized.

Change-Id: I4ee56a6b396734ce182bf75d38a6b318138d0dc5