Fix possible crash when typing search text.
When a character is typed into the search field, the app sends a message
that triggers the search after a predefined delay. If the search fragment
was detached during that delay period, the app would crash. This patch
will prevent any execution after receiving the message, if the fragment is
no longer attached.

Bug: T85429
Change-Id: I2cb8be69c776d32a6b29bfab68640a850f99cf8c