Don't show lead image if it's a .GIF
If the API returns a GIF image as the lead image for an article, then
simply don't show it.  GIF images are usually used for mathematical or
engineering diagrams, and don't look good as "lead" images. Moreover, GIFs
are very likely to be animated, which also isn't good (and doesn't work)
for a lead image.

Change-Id: Ie91c59403d9747dcb85e7d8b4fba1ff3973dc6ef