Add "?source=app" to the end of all share a fact URLs per Analytics.
Our EventLogging infrastructure captures the generation and sharing of facts,
but Analytics would like the ability to track how much the links are clicked,
and the best way to do that is to add a URL parameter tracking the source.
This patch adds such a parameter, conforming to the spec given to me by the
Analytics team.

Change-Id: Ie43d02653c00f40c4e1a95bcd7b05ac32f33da56