Improve list of media items shown in Gallery.
No longer including SVG and PNG items in the Gallery by default. This
effectively gets rid of most, if not all, irrelevant presentational images
(featured star, wikimedia logo, etc).

Of course, if the user explicitly clicks on an SVG image in a page, it
will be shown in the gallery, in addition to the other default items.

Bug: T86073
Change-Id: I6013ed21c73ce48544c836366adb3b96df4dedca