Combine title and full text search
No need to explicitly do an auto switch based on a minimum
number of title search results since it automatically does so
when the end of the list is reached, either by not having enough
results to fill the screen or by the user scrolling to the end.

Increased the batch size to 20 to avoid immediate full text
searches on large devices.

Change-Id: Ida613bd18716eef09149b41b0483127e77de07e2