Properly handle back button in long-press contexts.
This fixes the back-button behavior when inside the long-press context
in the History and SavedPages fragments.

It also fixes a crashing bug when long-pressing items in History or SavedPages
because the PageActivity was getting its ActionModes mixed up. We're now
assigning a "tag" to each different ActionMode, and the PageActivity now knows
what to do, based on the tag.

Bug: T87052
Change-Id: Ibbba1330eae958432923fd0baee6b49f33918231