Add script to find undocumented translations.
To my knowledge there's no convenient way to find out if a string is missing a
translation into a particular language. For us this means that it's not easy
to check all of our strings and make sure they have documentation for our
translators. This patch adds a Python script that, when run, will tell you
what messages are undocumented.

Change-Id: I33b51f314aea841dda9f11bbcaf22158ee73960b