More aggressively disable autocorrect on username fields
Although we had already set the fields to not give suggestions, certain
classes of Android devices, such as some Samsung devices, ignore the
instruction and display suggestions anyway. Even devices that do not
(like my stock Nexus 5) still reserve space for the suggestions that
don't even appear.

This patch marks username fields as visible password fields as a
workaround to both of the above.

Bug: T87208
Change-Id: Ib8edee8e49f4a2b248112ac1e541c9ece9e776e9