Enable search suggestions in prefix search task
Using list=search with srlimit of 1. Would have used srlimit of 0
if the API allowed it. This makes suggestions show up some of
the times.

Originally I added it to full text search but then the suggestions
comes too late. The first request is the prefix search. The
full text search comes as a second request.

Also refactored the SearchResult class to be used by both
prefix and full text searches.

Bug: T86982
Change-Id: I43a53b52069378af3cb08bb564b5f99a0511e5bd