Avoid NPE crash from onPrepareOptionsMenu
Maybe related to https://github.com/Prototik/HoloEverywhere/issues/577
'In onPrepareOptionsMenu, calling menu.findItem returns null when pressing "menu" key'

Also refactored the findItem calls since they are potentially expensive
and savePageMenu is used multiple times.

I assume that if the first menu item is not null the others would not be
null as well. If that's not the case we should get more reports like this
with newer releases. But it think that's unlikely.

Bug: T87109
Change-Id: Ic956ce57ea8a440c7d887978b43acdd432af92f8