Switch to using PreferenceFragment for settings.
As of API 11, the proper way to present lists of settings is by using a
PreferenceFragment instead of PreferenceActivity.  And because the activity
that contains the new Fragment inherits from ActionBarActivity, it indirectly
fixes the rendering issues that we saw with checkboxes.

Unfortunately API 10 does not support PreferenceFragment, so we have to keep
the old PreferenceActivity specifically for compatibility with API 10.

Also reverted the previous fix for the checkbox rendering issue.

Bug: T85809
Change-Id: I61d93eade8d0f059411713898b816b89bb32b7a8