Fix scrolling when tapping and dragging over lead image.
Added a custom OnClickListener (with coordinates) to the actual WebView.
Then, if we detect a click within the logical area of the lead image, then
we can assume that the user intended to click on the lead image. Otherwise
the touch/scroll/etc event will be passed on to the WebView.

Bug: T87128
Change-Id: Ie32524661c4d5408cc17489898074c5bceab33be