Share first paragraph when no text is selected
This functionality is accessible from the Share overflow menu.
If "Share as text" is selected just share the URL.

For when text is selected:
If less than 2 characters are selected also just share the URL.

Added TODO for NoTextSelectedShareAdapter because
right now it's not clear if we want to have event logged
when no text was selected.

Change-Id: I7298d08904713ef0f27d042fea31e61a7672dfef