Correct misleading comment in handleIntent method in PageActivity
This patch corrects a slightly misleading, but technically correct, comment in
the handleIntent method in PageActivity. The comment implies that the else
statement only executes when the app is handled an Intent it doesn't
recognise, which is not actually the case. Through doing some testing using
breakpoints, I've found out that this else statement is also triggered if the
user taps on the app from their home screen, which is far from out of the
ordinary!

This is a pretty subjective matter, so I intentionally made this change in a
separate commit from the other commit I just made that touches this area of
the code. Feel free to tell me I'm silly and that I should abandon this. :-)

Change-Id: If2b514d1f32a37a5baa46518251cc74c17bf726e