Fix ability to share text selection after screen rotation.
Refactored the way we initialize and clean up our ShareAdapters (both the
TextSelected and NoTextSelected adapters).
They were being instantiated in-place, and were creating Dialog objects
that weren't being destroyed properly upon screen rotation (leaking the
window).  This was also causing an imbalance in the adding and removing of
the Clipboard Listener that receives the shared text from the WebView,
which was preventing the text from being shared after rotation.

I modified the ShareAdapters to inherit from the base ShareHandler, which
will maintain a reference to the Dialog that it instantiates, so that it
can be explicitly dismissed when necessary.  And the Clipboard listener
will now persist for the lifetime of the Activity.

Bug: T89979
Change-Id: Ib081a48ef342c0844fa996c94314d41b6ff9d298