Use icons in Gallery license information.
Using logic loosely based on the current MultimediaViewer extension's
handling of license information.

If a media item explicitly licensed under CC (any variation), then show
the CC icon. If an item is explicitly licensed as PD, then show the PD
icon. In all other cases, show the "cite" icon.

Clicking on the icon will launch the associated license URL in the
external browser.

Long-pressing the icon will pop up a Toast with the UsageTerms for the
item.

Change-Id: I621b87b2ae1cc89189c7cf7d8e9d1cf030a881c2