Don't display last modified message on main page
The last modified time is always wrong, because the contents of the main page
are contained inside another template. This patch simply hides the message. An
alternative would've been to hack it to always say that it was modified today,
but that could be wrong on wikis where the main page content isn't updated
daily.

This patch also adds isFilePage() and isMainPage() methods to the Page class
for convenience. It was quite clunky to remember that you needed to
getPageProperties().isMainPage() and getTitle.isFilePage(), so this alleviates
you having to do that if you have the Page object. A subsequent patch can
migrate places where the old way is used to the more convenient way.

Bug: T78005
Change-Id: Ib9e9ac8b99755785cd34bed535e23d5df4db69bc