ShareAFact: EL update
Also log share as text.
Added "sharemode" enum and new actions "sharetap" and "abandoned".

The new "failure" action is not feasible on Android for the same reason
as "share" is not.

Restrict the length of text to 99  (for now) to avoid events being dropped.
Dan, feel free to change as needed.

Moved setting and checking of currentActionMode around to avoid logging
the highlight event twice.

Change-Id: I5f0a88c33e3ef31f6d5e6e5e102468be5a6f72d1