Unregister clipboard listener in onDestroy instead of onStop.
The onStop event is too "early" for unregistering the clipboard listener.
It's possible to go to another activity (e.g. gallery), then come back,
and suddenly be unable to share highlighted text. This is because the
activity was stopped (onStop) and then resumed.

The onDestroy event is called when the activity will definitely not be
resumed. Dismissing the sharing dialogs inside this event is OK, too.

Change-Id: I89d2e0ded02929434dc3d55cb34afbaccb77c901