Use designtime layout attributes
The namespace prefixed with "tools" is stripped out by the compiler,
and only used for IDE and other tool support. So it's safe to leave this in
since it doesn't get included into the apk.

This is great for overriding or setting attributes just for the
IDE's layout preview/designer.

More at http://tools.android.com/tips/layout-designtime-attributes

Change-Id: Id942a78801648196b689d7571c4c6a34c11e46ae