Just log the read next items actually shown
Before it was logging up to 11 items when it only showed 1
for the new Read Next functionality.

Now callers of SuggestionsTask specify both the number
of items requested and the max number of items shown.

For the case where a lead image is required just one extra
item doesn't cut if since there could be items without
lead images in the result returned by the API.

Also fixed the test.

Change-Id: I26be484d9f98d51dadf96de5c4736fa292f387c2