Update launcher icon.
Also no longer generate the icon from SVG, since the new design of the
icon (done in Adobe Illustrator) uses SVG elements that don't seem to be
compatible with rsvg.

For the time being, the policy shall be to receive a 512x512 icon from
designers, and size it down to the desired resolutions.

Added 512x512 icons in a separate folder. May be beneficial to add a
script that creates the various sizes based on the originals, in the
future.

Change-Id: Ic943f410369fd0ff1e72296b702cdadb470aa38f