Fix list item background in 2.3.
Under 2.3, the list items (search results, history, etc) were showing up
with a solid blue (selected) background. It looks like the state_activated
property is having an effect on the background, even though it's not
supposed to apply to 2.3.  I've split off the list background selector for
API 11+, and modified the one for API 10 to no longer have
state_activated.

Change-Id: Ie3e29c151f9989e8c23ace09a188a84d00f660ca