Resolve redirects when doing prefixsearch.
Our search experience provides the user with prefix-search and full-text
search results. However, in the case of prefix-search results, if the
result is a Redirect page, it may be more useful to the user to see the
*target* of the redirect, rather than the redirect name itself.  In
addition, the redirect (source) page doesn't come with a thumbnail or
pageterms, which is even less useful to the user.

This patch enables automatic resolving of redirects for prefix-search,
whereby the target of the redirect will be shown, complete with thumbnail
and description. This is accomplished in a slightly roundabout way:
- The search API allows us to specify "redirects=true" to automatically
  resolve redirects. Simple enough...
- However, when we receive the results, any result that is a redirect
  target will no longer have the "index" property, which we use to sort
  the results in our list.
- However, the query response also contains a separate "redirects" array
  that contains a list of all the redirected results, apparently in the
  correct order.
- So first, I take all the results that have an "index" property, and
  make a list of the "holes" in the indices. I then fill the holes with
  the results that don't have an "index" property, in the order that they
  appear in the "redirects" list!

This can be vastly simplified, of course, if the API includes the "index"
property with the redirected results. However, this may be difficult to do
on the API side, so this will have to be the solution for now.

Change-Id: I65e281fa83071a4dbf288ecdd7d3bf5ee39a0aee