Don't crash when the AbuseFilter emits an arbitrary error code.
The abuse filter unfortunately allows admins to specify filters that
emit arbitrary codes. This is a nightmare to handle properly.

Right now, our lack of proper handling for these arbitrary codes is
causing the app to crash if one is received from the API. This patch
changes that behaviour to simply treat arbitrary codes as errors. That
sucks, but it's better than crashing.

Bug: T93539
Change-Id: Ib839d4104f4a51b344dc7b00faf7a8bcad069fe5