ShareAFact: use static image for English WP wordmark
and for languages that have the same text for the wordmark.
Other languages still have the wordmark drawn using text,
meaning the type font is not correct.

Aligned bottom of wordmark with bottom of CC-BY-SA icons.
Adjusted the text variant to match better the image variant
(smaller font, stretch a bit horizontally).

Started L10nUtils, which I hope we can move more of the L10N
specific util methods from Utils to this new class.

Bug: T94760
Change-Id: I4b996a0908313a8eaad37eccb0c15d0d3ee9b320