Make enter key submit request when the user is filling in a captcha.
Since the captcha form requires only a single input (i.e. the captcha),
this patch makes the enter key on the keyboard submit the form rather
than simply collapsing the keyboard.

This patch also slightly tweaks the ordering of the XML for the captcha
group, so that the android:id is at the top, which is our prevailing
style.

Bug: T95145
Change-Id: I68d37d8dd28a2a9c4741633cb646259b07c7cb6b